requirements.txt
fastapi
uvicorn[standard]
websockets
jinja2
pyjwt
numpy
joblib
ultralytics
opencv-python
xgboost
scikit-learn
fpdf
python-dotenv
httpx
chromadb
sentence-transformers
pypdf
llama-cpp-python
